move safe outputs logic out
