+++ pkg/workflow/compile_create_code_scanning_alert.go
@@ -100,3 +100,38 @@ func (c *Compiler) buildCreateOutputCodeScanningAlertJob(data *WorkflowData, mai
 
 	return job, nil
 }
+
+// parseCodeScanningAlertsConfig handles create-code-scanning-alert configuration
+func (c *Compiler) parseCodeScanningAlertsConfig(outputMap map[string]any) *CreateCodeScanningAlertsConfig {
+	if _, exists := outputMap["create-code-scanning-alert"]; !exists {
+		return nil
+	}
+
+	configData := outputMap["create-code-scanning-alert"]
+	securityReportsConfig := &CreateCodeScanningAlertsConfig{Max: 0} // Default max is 0 (unlimited)
+
+	if configMap, ok := configData.(map[string]any); ok {
+		// Parse max
+		if max, exists := configMap["max"]; exists {
+			if maxInt, ok := parseIntValue(max); ok {
+				securityReportsConfig.Max = maxInt
+			}
+		}
+
+		// Parse driver
+		if driver, exists := configMap["driver"]; exists {
+			if driverStr, ok := driver.(string); ok {
+				securityReportsConfig.Driver = driverStr
+			}
+		}
+
+		// Parse github-token
+		if githubToken, exists := configMap["github-token"]; exists {
+			if githubTokenStr, ok := githubToken.(string); ok {
+				securityReportsConfig.GitHubToken = githubTokenStr
+			}
+		}
+	}
+
+	return securityReportsConfig
+}
