+++ pkg/workflow/compiler.go
@@ -3420,32 +3420,7 @@ func (c *Compiler) extractSafeOutputsConfig(frontmatter map[string]any) *SafeOut
 	return config
 }
 
-// addCustomSafeOutputEnvVars adds custom environment variables to safe output job steps
-func (c *Compiler) addCustomSafeOutputEnvVars(steps *[]string, data *WorkflowData) {
-	if data.SafeOutputs != nil && len(data.SafeOutputs.Env) > 0 {
-		for key, value := range data.SafeOutputs.Env {
-			*steps = append(*steps, fmt.Sprintf("          %s: %s\n", key, value))
-		}
-	}
-}
 
-// addSafeOutputGitHubToken adds github-token to the with section of github-script actions
-func (c *Compiler) addSafeOutputGitHubToken(steps *[]string, data *WorkflowData) {
-	if data.SafeOutputs != nil && data.SafeOutputs.GitHubToken != "" {
-		*steps = append(*steps, fmt.Sprintf("          github-token: %s\n", data.SafeOutputs.GitHubToken))
-	}
-}
-
-// addSafeOutputGitHubTokenForConfig adds github-token to the with section, preferring per-config token over global
-func (c *Compiler) addSafeOutputGitHubTokenForConfig(steps *[]string, data *WorkflowData, configToken string) {
-	token := configToken
-	if token == "" && data.SafeOutputs != nil {
-		token = data.SafeOutputs.GitHubToken
-	}
-	if token != "" {
-		*steps = append(*steps, fmt.Sprintf("          github-token: %s\n", token))
-	}
-}
 
 // extractCacheMemoryConfig extracts cache-memory configuration from tools section
 func (c *Compiler) extractCacheMemoryConfig(tools map[string]any) *CacheMemoryConfig {
@@ -3504,426 +3479,25 @@ func (c *Compiler) extractCacheMemoryConfig(tools map[string]any) *CacheMemoryCo
 	return nil
 }
 
-// parseIssuesConfig handles create-issue configuration
-func (c *Compiler) parseIssuesConfig(outputMap map[string]any) *CreateIssuesConfig {
-	if configData, exists := outputMap["create-issue"]; exists {
-		issuesConfig := &CreateIssuesConfig{Max: 1} // Default max is 1
 
-		if configMap, ok := configData.(map[string]any); ok {
-			// Parse title-prefix
-			if titlePrefix, exists := configMap["title-prefix"]; exists {
-				if titlePrefixStr, ok := titlePrefix.(string); ok {
-					issuesConfig.TitlePrefix = titlePrefixStr
-				}
-			}
 
-			// Parse labels
-			if labels, exists := configMap["labels"]; exists {
-				if labelsArray, ok := labels.([]any); ok {
-					var labelStrings []string
-					for _, label := range labelsArray {
-						if labelStr, ok := label.(string); ok {
-							labelStrings = append(labelStrings, labelStr)
-						}
-					}
-					issuesConfig.Labels = labelStrings
-				}
-			}
 
-			// Parse max
-			if max, exists := configMap["max"]; exists {
-				if maxInt, ok := c.parseIntValue(max); ok {
-					issuesConfig.Max = maxInt
-				}
-			}
 
-			// Parse github-token
-			if githubToken, exists := configMap["github-token"]; exists {
-				if githubTokenStr, ok := githubToken.(string); ok {
-					issuesConfig.GitHubToken = githubTokenStr
-				}
-			}
-		}
 
-		return issuesConfig
-	}
 
-	return nil
-}
 
-// parseDiscussionsConfig handles create-discussion configuration
-func (c *Compiler) parseDiscussionsConfig(outputMap map[string]any) *CreateDiscussionsConfig {
-	if configData, exists := outputMap["create-discussion"]; exists {
-		discussionsConfig := &CreateDiscussionsConfig{Max: 1} // Default max is 1
 
-		if configMap, ok := configData.(map[string]any); ok {
-			// Parse title-prefix
-			if titlePrefix, exists := configMap["title-prefix"]; exists {
-				if titlePrefixStr, ok := titlePrefix.(string); ok {
-					discussionsConfig.TitlePrefix = titlePrefixStr
-				}
-			}
 
-			// Parse category-id
-			if categoryId, exists := configMap["category-id"]; exists {
-				if categoryIdStr, ok := categoryId.(string); ok {
-					discussionsConfig.CategoryId = categoryIdStr
-				}
-			}
 
-			// Parse max
-			if max, exists := configMap["max"]; exists {
-				if maxInt, ok := c.parseIntValue(max); ok {
-					discussionsConfig.Max = maxInt
-				}
-			}
 
-			// Parse github-token
-			if githubToken, exists := configMap["github-token"]; exists {
-				if githubTokenStr, ok := githubToken.(string); ok {
-					discussionsConfig.GitHubToken = githubTokenStr
-				}
-			}
-		}
 
-		return discussionsConfig
-	}
 
-	return nil
-}
 
-// parseCommentsConfig handles add-comment configuration
-func (c *Compiler) parseCommentsConfig(outputMap map[string]any) *AddCommentsConfig {
-	if configData, exists := outputMap["add-comment"]; exists {
-		commentsConfig := &AddCommentsConfig{Max: 1} // Default max is 1
 
-		if configMap, ok := configData.(map[string]any); ok {
-			// Parse max
-			if max, exists := configMap["max"]; exists {
-				if maxInt, ok := c.parseIntValue(max); ok {
-					commentsConfig.Max = maxInt
-				}
-			}
 
-			// Parse target
-			if target, exists := configMap["target"]; exists {
-				if targetStr, ok := target.(string); ok {
-					commentsConfig.Target = targetStr
-				}
-			}
 
-			// Parse github-token
-			if githubToken, exists := configMap["github-token"]; exists {
-				if githubTokenStr, ok := githubToken.(string); ok {
-					commentsConfig.GitHubToken = githubTokenStr
-				}
-			}
-		}
 
-		return commentsConfig
-	}
 
-	return nil
-}
-
-// parsePullRequestsConfig handles only create-pull-request (singular) configuration
-func (c *Compiler) parsePullRequestsConfig(outputMap map[string]any) *CreatePullRequestsConfig {
-	// Check for singular form only
-	if _, exists := outputMap["create-pull-request"]; !exists {
-		return nil
-	}
-
-	configData := outputMap["create-pull-request"]
-	pullRequestsConfig := &CreatePullRequestsConfig{Max: 1} // Always max 1 for pull requests
-
-	if configMap, ok := configData.(map[string]any); ok {
-		// Parse title-prefix
-		if titlePrefix, exists := configMap["title-prefix"]; exists {
-			if titlePrefixStr, ok := titlePrefix.(string); ok {
-				pullRequestsConfig.TitlePrefix = titlePrefixStr
-			}
-		}
-
-		// Parse labels
-		if labels, exists := configMap["labels"]; exists {
-			if labelsArray, ok := labels.([]any); ok {
-				var labelStrings []string
-				for _, label := range labelsArray {
-					if labelStr, ok := label.(string); ok {
-						labelStrings = append(labelStrings, labelStr)
-					}
-				}
-				pullRequestsConfig.Labels = labelStrings
-			}
-		}
-
-		// Parse draft
-		if draft, exists := configMap["draft"]; exists {
-			if draftBool, ok := draft.(bool); ok {
-				pullRequestsConfig.Draft = &draftBool
-			}
-		}
-
-		// Parse if-no-changes
-		if ifNoChanges, exists := configMap["if-no-changes"]; exists {
-			if ifNoChangesStr, ok := ifNoChanges.(string); ok {
-				pullRequestsConfig.IfNoChanges = ifNoChangesStr
-			}
-		}
-
-		// Parse github-token
-		if githubToken, exists := configMap["github-token"]; exists {
-			if githubTokenStr, ok := githubToken.(string); ok {
-				pullRequestsConfig.GitHubToken = githubTokenStr
-			}
-		}
-
-		// Note: max parameter is not supported for pull requests (always limited to 1)
-		// If max is specified, it will be ignored as pull requests are singular only
-	}
-
-	return pullRequestsConfig
-}
-
-// parsePullRequestReviewCommentsConfig handles create-pull-request-review-comment configuration
-func (c *Compiler) parsePullRequestReviewCommentsConfig(outputMap map[string]any) *CreatePullRequestReviewCommentsConfig {
-	if _, exists := outputMap["create-pull-request-review-comment"]; !exists {
-		return nil
-	}
-
-	configData := outputMap["create-pull-request-review-comment"]
-	prReviewCommentsConfig := &CreatePullRequestReviewCommentsConfig{Max: 10, Side: "RIGHT"} // Default max is 10, side is RIGHT
-
-	if configMap, ok := configData.(map[string]any); ok {
-		// Parse max
-		if max, exists := configMap["max"]; exists {
-			if maxInt, ok := c.parseIntValue(max); ok {
-				prReviewCommentsConfig.Max = maxInt
-			}
-		}
-
-		// Parse side
-		if side, exists := configMap["side"]; exists {
-			if sideStr, ok := side.(string); ok {
-				// Validate side value
-				if sideStr == "LEFT" || sideStr == "RIGHT" {
-					prReviewCommentsConfig.Side = sideStr
-				}
-			}
-		}
-
-		// Parse github-token
-		if githubToken, exists := configMap["github-token"]; exists {
-			if githubTokenStr, ok := githubToken.(string); ok {
-				prReviewCommentsConfig.GitHubToken = githubTokenStr
-			}
-		}
-	}
-
-	return prReviewCommentsConfig
-}
-
-// parseCodeScanningAlertsConfig handles create-code-scanning-alert configuration
-func (c *Compiler) parseCodeScanningAlertsConfig(outputMap map[string]any) *CreateCodeScanningAlertsConfig {
-	if _, exists := outputMap["create-code-scanning-alert"]; !exists {
-		return nil
-	}
-
-	configData := outputMap["create-code-scanning-alert"]
-	securityReportsConfig := &CreateCodeScanningAlertsConfig{Max: 0} // Default max is 0 (unlimited)
-
-	if configMap, ok := configData.(map[string]any); ok {
-		// Parse max
-		if max, exists := configMap["max"]; exists {
-			if maxInt, ok := c.parseIntValue(max); ok {
-				securityReportsConfig.Max = maxInt
-			}
-		}
-
-		// Parse driver
-		if driver, exists := configMap["driver"]; exists {
-			if driverStr, ok := driver.(string); ok {
-				securityReportsConfig.Driver = driverStr
-			}
-		}
-
-		// Parse github-token
-		if githubToken, exists := configMap["github-token"]; exists {
-			if githubTokenStr, ok := githubToken.(string); ok {
-				securityReportsConfig.GitHubToken = githubTokenStr
-			}
-		}
-	}
-
-	return securityReportsConfig
-}
-
-// parseIntValue safely parses various numeric types to int
-func (c *Compiler) parseIntValue(value any) (int, bool) {
-	switch v := value.(type) {
-	case int:
-		return v, true
-	case int64:
-		return int(v), true
-	case uint64:
-		return int(v), true
-	case float64:
-		return int(v), true
-	default:
-		return 0, false
-	}
-}
-
-// parseUpdateIssuesConfig handles update-issue configuration
-func (c *Compiler) parseUpdateIssuesConfig(outputMap map[string]any) *UpdateIssuesConfig {
-	if configData, exists := outputMap["update-issue"]; exists {
-		updateIssuesConfig := &UpdateIssuesConfig{Max: 1} // Default max is 1
-
-		if configMap, ok := configData.(map[string]any); ok {
-			// Parse max
-			if max, exists := configMap["max"]; exists {
-				if maxInt, ok := c.parseIntValue(max); ok {
-					updateIssuesConfig.Max = maxInt
-				}
-			}
-
-			// Parse target
-			if target, exists := configMap["target"]; exists {
-				if targetStr, ok := target.(string); ok {
-					updateIssuesConfig.Target = targetStr
-				}
-			}
-
-			// Parse status - presence of the key (even if nil/empty) indicates field can be updated
-			if _, exists := configMap["status"]; exists {
-				// If the key exists, it means we can update the status
-				// We don't care about the value - just that the key is present
-				updateIssuesConfig.Status = new(bool) // Allocate a new bool pointer (defaults to false)
-			}
-
-			// Parse title - presence of the key (even if nil/empty) indicates field can be updated
-			if _, exists := configMap["title"]; exists {
-				updateIssuesConfig.Title = new(bool)
-			}
-
-			// Parse body - presence of the key (even if nil/empty) indicates field can be updated
-			if _, exists := configMap["body"]; exists {
-				updateIssuesConfig.Body = new(bool)
-			}
-
-			// Parse github-token
-			if githubToken, exists := configMap["github-token"]; exists {
-				if githubTokenStr, ok := githubToken.(string); ok {
-					updateIssuesConfig.GitHubToken = githubTokenStr
-				}
-			}
-		}
-
-		return updateIssuesConfig
-	}
-
-	return nil
-}
-
-// parsePushToPullRequestBranchConfig handles push-to-pr-branch configuration
-func (c *Compiler) parsePushToPullRequestBranchConfig(outputMap map[string]any) *PushToPullRequestBranchConfig {
-	if configData, exists := outputMap["push-to-pr-branch"]; exists {
-		pushToBranchConfig := &PushToPullRequestBranchConfig{
-			IfNoChanges: "warn", // Default behavior: warn when no changes
-		}
-
-		// Handle the case where configData is nil (push-to-pr-branch: with no value)
-		if configData == nil {
-			return pushToBranchConfig
-		}
-
-		if configMap, ok := configData.(map[string]any); ok {
-			// Parse target (optional, similar to add-comment)
-			if target, exists := configMap["target"]; exists {
-				if targetStr, ok := target.(string); ok {
-					pushToBranchConfig.Target = targetStr
-				}
-			}
-
-			// Parse if-no-changes (optional, defaults to "warn")
-			if ifNoChanges, exists := configMap["if-no-changes"]; exists {
-				if ifNoChangesStr, ok := ifNoChanges.(string); ok {
-					// Validate the value
-					switch ifNoChangesStr {
-					case "warn", "error", "ignore":
-						pushToBranchConfig.IfNoChanges = ifNoChangesStr
-					default:
-						// Invalid value, use default and log warning
-						if c.verbose {
-							fmt.Printf("Warning: invalid if-no-changes value '%s', using default 'warn'\n", ifNoChangesStr)
-						}
-						pushToBranchConfig.IfNoChanges = "warn"
-					}
-				}
-			}
-
-			// Parse github-token
-			if githubToken, exists := configMap["github-token"]; exists {
-				if githubTokenStr, ok := githubToken.(string); ok {
-					pushToBranchConfig.GitHubToken = githubTokenStr
-				}
-			}
-		}
-
-		return pushToBranchConfig
-	}
-
-	return nil
-}
-
-// parseMissingToolConfig handles missing-tool configuration
-func (c *Compiler) parseMissingToolConfig(outputMap map[string]any) *MissingToolConfig {
-	if configData, exists := outputMap["missing-tool"]; exists {
-		missingToolConfig := &MissingToolConfig{} // Default: no max limit
-
-		// Handle the case where configData is nil (missing-tool: with no value)
-		if configData == nil {
-			return missingToolConfig
-		}
-
-		if configMap, ok := configData.(map[string]any); ok {
-			// Parse max (optional)
-			if max, exists := configMap["max"]; exists {
-				// Handle different numeric types that YAML parsers might return
-				var maxInt int
-				var validMax bool
-				switch v := max.(type) {
-				case int:
-					maxInt = v
-					validMax = true
-				case int64:
-					maxInt = int(v)
-					validMax = true
-				case uint64:
-					maxInt = int(v)
-					validMax = true
-				case float64:
-					maxInt = int(v)
-					validMax = true
-				}
-				if validMax {
-					missingToolConfig.Max = maxInt
-				}
-			}
-
-			// Parse github-token
-			if githubToken, exists := configMap["github-token"]; exists {
-				if githubTokenStr, ok := githubToken.(string); ok {
-					missingToolConfig.GitHubToken = githubTokenStr
-				}
-			}
-		}
-
-		return missingToolConfig
-	}
-
-	return nil
-}
 
 // buildCustomJobs creates custom jobs defined in the frontmatter jobs section
 func (c *Compiler) buildCustomJobs(data *WorkflowData) error {
